+++ .gitignore
@@ -32,7 +32,7 @@ index.html.md
 libs/pubsub/query/fuzz_test/output
 profile\.out
 remote_dump
-rpc/test/.tendermint
+node/rpc/test/.tendermint
 scripts/cutWALUntil/cutWALUntil
 scripts/wal2json/wal2json
 shunit2
