+++ node/config.go
@@ -2,14 +2,14 @@ package node
 
 import (
 	"github.com/celestiaorg/celestia-node/node/p2p"
-	"github.com/celestiaorg/celestia-node/rpc"
+	rpc2 "github.com/celestiaorg/celestia-node/node/rpc"
 )
 
 // Config is main configuration structure for a Node.
 // It combines configuration units for all Node subsystems.
 type Config struct {
 	P2P *p2p.Config
-	RPC *rpc.Config
+	RPC *rpc2.Config
 }
 
 // DefaultConfig provides a default Node Config.
