+++ node/full_test.go
@@ -2,6 +2,7 @@ package node
 
 import (
 	"context"
+	rpc2 "github.com/celestiaorg/celestia-node/node/rpc"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -11,7 +12,6 @@ import (
 	core_node "github.com/celestiaorg/celestia-core/node"
 	rpctest "github.com/celestiaorg/celestia-core/rpc/test"
 	"github.com/celestiaorg/celestia-node/node/p2p"
-	"github.com/celestiaorg/celestia-node/rpc"
 )
 
 func TestNewFull(t *testing.T) {
@@ -25,7 +25,7 @@ func TestNewFull(t *testing.T) {
 
 	nd, err := NewFull(&Config{
 		P2P: &p2p.Config{},
-		RPC: &rpc.Config{
+		RPC: &rpc2.Config{
 			Protocol:   protocol,
 			RemoteAddr: ip,
 		},
@@ -45,7 +45,7 @@ func TestFullLifecycle(t *testing.T) {
 
 	node, err := NewFull(&Config{
 		P2P: &p2p.Config{},
-		RPC: &rpc.Config{
+		RPC: &rpc2.Config{
 			Protocol:   protocol,
 			RemoteAddr: ip,
 		},
