+++ node/rpc/config.go
@@ -1,6 +1,6 @@
 package rpc
 
-// RPCConfig represents the configuration for the RPC client
+// Config represents the configuration for the RPC client
 // used for retrieving information from Celestia Core nodes // TODO @renaynay: update this post-devnet
 type Config struct {
 	Protocol   string
